requirements.txt
amqp==5.3.1
apturl==0.5.2
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
bcrypt==3.2.0
billiard==4.2.1
black==25.1.0
blinker==1.9.0
Brlapi==0.8.3
celery==5.5.3
certifi==2020.6.20
chardet==4.0.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.4
command-not-found==0.3
coverage==7.4.3
cron-descriptor==1.4.5
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
decorator==5.2.1
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1.7
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-filter==25.1
django-redis==5.4.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
duplicity==0.8.21
exceptiongroup==1.2.0
executing==2.2.0
fasteners==0.14.1
flake8==7.1.2
Flask==3.1.0
future==0.18.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
iniconfig==2.0.0
ipython==8.34.0
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
keyring==23.5.0
kombu==5.5.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
monotonic==1.6
more-itertools==8.10.0
mypy_extensions==1.1.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
orjson==3.9.15
packaging==24.0
paramiko==2.9.3
parso==0.8.4
pathlib==1.0.1
pathspec==0.12.1
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.3.8
pluggy==1.4.0
prompt_toolkit==3.0.50
protobuf==3.12.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycairo==1.20.1
pycodestyle==2.12.1
pycups==2.0.1
pyflakes==3.2.0
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==8.1.1
pytest-cov==4.1.0
python-apt==2.4.0+ubuntu4
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==1.1.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==6.1.0
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.5.3
stack-data==0.6.3
stripe==12.2.0
systemd-python==234
tomli==2.0.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.2.0
urllib3==1.26.5
usb-creator==0.3.7
vine==5.1.0
wadllib==1.3.6
wcwidth==0.2.13
Werkzeug==3.1.3
xdg==5
xkit==0.0.0
zipp==1.0.0
